% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_congress.R
\name{calc_congress}
\alias{calc_congress}
\title{Calculate congress corresponding to a year}
\usage{
calc_congress(year)
}
\arguments{
\item{year}{integer year}
}
\value{
integer congress
}
\description{
Simple conversion between year and congress
}
