% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_switch_rate.R
\name{calc_switch_rate}
\alias{calc_switch_rate}
\title{TITLE}
\usage{
calc_switch_rate(rc)
}
\arguments{
\item{rc}{xxx}

\item{DT}{xxx}

\item{tvals}{xxx}

\item{type}{xxx}
}
\value{
xxx
}
\description{
DETAIL
}
