% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_party_vote_rates.R
\name{calc_vote_with_party_rates}
\alias{calc_vote_with_party_rates}
\title{Calculate `vote with party' rates for legislators}
\usage{
calc_vote_with_party_rates(congress_index, roll_call_list)
}
\arguments{
\item{congress_index}{integer from 1 to 20}

\item{roll_call_list}{list of roll call objects with votes coded as 1 and -1}
}
\value{
data.table with congress, icpsrLegis, and party vote rate
}
\description{
`Vote with party` rates are the fraction of times a legislator votes with her
party on all votes.
}
