% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_year.R
\name{calc_year}
\alias{calc_year}
\title{Calculate year corresponding to a congress}
\usage{
calc_year(congress)
}
\arguments{
\item{congress}{integer congress}
}
\value{
integer year
}
\description{
Simple conversion between congress and year
}

