% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_year_elected.R
\name{calc_year_elected}
\alias{calc_year_elected}
\title{Calculate election year corresponding to a congress}
\usage{
calc_year_elected(congress, class)
}
\arguments{
\item{congress}{integer congress}

\item{class}{integer 1, 2, or 3; see Detail}
}
\value{
integer year
}
\description{
Calculate year of most previous regular election based on congress and class
class 1: up for reelection in 1964, 1970, 1976, 1982, 1988, 1994, 2000, 2006,
 2012; congresses in seq(94, 112, 3)
class 2: up for reelection in 1966, 1972, 1978, 1984, 1990, 1996, 2002, 2008,
 2014; congresses in seq(95, 113, 3)
class 3: up for reelection in 1968, 1974, 1980, 1986, 1992, 1998, 2004, 2010,
 2016; congresses in seq(93, 114, 3)
}
