% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{code_party_calls}
\alias{code_party_calls}
\title{Run the party calls classifier}
\usage{
code_party_calls(rc)
}
\arguments{
\item{rc}{a rollcall object}
}
\value{
rollcall object with record of classification algorithm and
list of classified party calls
}
\description{
Use the iterative algorithm that (1) predicts ideal points based on last
iteration's non-party calls, (2) runs regressions of roll call vote on
ideal points and party, (3) classify new iteration of non-party calls as
votes for which the p value on party is > .05.
}

