% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{code_party_calls_1step}
\alias{code_party_calls_1step}
\title{Run one step of the party calls algorithm}
\usage{
code_party_calls_1step(rc, DT, noncalls, return_pvals, use_brglm)
}
\arguments{
\item{rc}{a rollcall object}

\item{DT}{data.table with votes and party indicators}

\item{noncalls}{indices for non-party calls from last run}

\item{return_pvals}{logical for whether to return pvals or tvals}

\item{use_brglm}{logical for whether to use the bias reduced logit}
}
\value{
vector of indices for non-party calls
}
\description{
To be used inside a call to code_party_calls
}

