% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_party_calls_by_congress_number.R
\name{code_party_calls_by_congress_number}
\alias{code_party_calls_by_congress_number}
\title{TITLE}
\usage{
code_party_calls_by_congress_number(congress_number, chamber = "house", ...)
}
\arguments{
\item{congress_number}{integer values for the congress to analyze}

\item{chamber}{character string, house or senate}

\item{...}{arguments to pass to \code{code_party_calls}}
}
\value{
xxx
}
\description{
DETAIL
}
