% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_party_call_coding.R
\name{get_party_call_coding}
\alias{get_party_call_coding}
\title{Get party call coding based on classifier results}
\usage{
get_party_call_coding(rc, n_iterations)
}
\arguments{
\item{rc}{rollcall object, includes record_of_coding}

\item{n_iterations}{number of classifier iterations to include in coding
of roll call votes}
}
\value{
data.table with two columns, voteno and coding
}
\description{
Examines the last n_iterations of classifier iterations and codes any
vote that was always a party call as a party call, any vote that was always
a noncall as a noncall, and the remainder as gray votes.
}

