% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_party_call_coding.R
\name{get_party_calls}
\alias{get_party_calls}
\title{Retrieve party calls}
\usage{
get_party_calls(rc, n_iterations = 5)
}
\arguments{
\item{rc}{rollcall object, includes record_of_coding}

\item{n_iterations}{number of classifier iterations to include in coding
of roll call votes}
}
\value{
character string of vote IDs, based on KH Data index
}
\description{
Examines the last n_iterations of classifier iterations and codes any
vote that was always a party call as a party call
}

