% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_party_vote_rates.R
\name{identify_party_votes}
\alias{identify_party_votes}
\title{Identify party votes}
\usage{
identify_party_votes(congress_index, roll_call_list)
}
\arguments{
\item{congress_index}{integer from 1 to 20}

\item{roll_call_list}{list of roll call objects with votes coded as 1 and -1}
}
\value{
integer vector of votes in the roll call object that are party votes
}
\description{
Party votes are those that pit majorities of the two parties against each
other.
}
