% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_member_year_data.R
\name{make_member_year_data}
\alias{make_member_year_data}
\title{Make dataset at level of MC/year}
\usage{
make_member_year_data(congress, roll_calls_object_list, chamber = "house")
}
\arguments{
\item{congress}{congress ID number}

\item{roll_calls_object_list}{rc a list of rollcall objects with classified
party calls}
}
\value{
data.table with party-free ideal points
}
\description{
For a single congress, assemble data for analysis
}

