% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-functions.R
\name{robust_vcov}
\alias{robust_vcov}
\title{TITLE}
\usage{
robust_vcov(model, cluster1 = NULL, cluster2 = NULL)
}
\arguments{
\item{model}{DESC}

\item{cluster1}{DESC}

\item{cluster2}{DESC}
}
\value{
DESC
}
\description{
DESC
}

