% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{symdiff}
\alias{symdiff}
\title{Find symmetric difference}
\usage{
symdiff(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
vector of elements in either x or y but not both
}
\description{
The symmetric difference is defined as the set difference between the
union of two sets and the intersection of those sets.
}

