% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_rollcall.R
\name{test_rollcall}
\alias{test_rollcall}
\title{Regress a single roll call on party indicator and ideal points}
\usage{
test_rollcall(.SD, type = c("brglm", "lm", "glm"))
}
\arguments{
\item{.SD}{subset of a data.table of roll call votes, with a column for party
labels}

\item{type}{character string, one of brglm, glm, or lm; which function
to use for roll call-by-roll call regression}
}
\value{
list of coefficient, standard error, t value, and p value for the
coefficient on party
}
\description{
To be used inside a call to code_party_calls_1step. This function is used to
perform a number of tasks in the purpose of determining the influence of the
party on a given vote. This is accomplished by regressing votes on members'
party affiliation and ideology. Before this happens, however, all votes which
either received only `yeas' and `nays' are assigned non-values in place of
estimates. Further, votes which strictly follow party lines are assigned
estimates which guarantee they will be coded as party calls under any model
specification to avoid separation in the model. Finally, the votes left are
either put through a bias-reduced logit or ordinary least squares model as
per user selected parameters. The default setting of the function is to use
the bias-reduced logit.
Internal function for code_party_calls, only used if
use_classification_distance is FALSE.
}

