% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergent-validity-functions.R
\name{test_rollcall2}
\alias{test_rollcall2}
\title{TITLE}
\usage{
test_rollcall2(.SD, type = c("brglm", "lm", "glm"))
}
\arguments{
\item{.SD}{DESC}

\item{type}{DESC}
}
\value{
DESC
}
\description{
DETAIL
}
