% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_party_calls_1step.R
\name{code_party_calls_1step}
\alias{code_party_calls_1step}
\title{Run one step of the party calls algorithm}
\usage{
code_party_calls_1step(rc, DT, votes_for_ideal_point_estimation, return_pvals,
  type = c("brglm", "lm", "glm"))
}
\arguments{
\item{rc}{a rollcall object}

\item{DT}{data.table with votes and party indicators}

\item{votes_for_ideal_point_estimation}{indices for votes to use in ideal
point estimation}

\item{return_pvals}{logical for whether to return pvals or tvals}

\item{type}{character string, one of brglm, glm, or lm; which function
to use for roll call-by-roll call regression}
}
\value{
list of pvals or tvals and ideal points
}
\description{
To be used inside a call to code_party_calls
}

