% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_classification}
\alias{test_classification}
\title{TITLE}
\usage{
test_classification(rc, DT, noncalls, tvals, type)
}
\arguments{
\item{rc}{xxx}

\item{DT}{xxx}

\item{tvals}{xxx}

\item{type}{xxx}
}
\value{
xxx
}
\description{
DETAIL
}

