% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_rollcall}
\alias{test_rollcall}
\title{Regress a single roll call on party indicator and ideal points}
\usage{
test_rollcall(.SD)
}
\arguments{
\item{.SD}{subset of a data.table of roll call votes, with a column for party
labels}
}
\value{
list of coefficient, standard error, t value, and p value for the
coefficient on party
}
\description{
To be used inside a call to code_party_calls_1step
}

